+++ Dockerfile
@@ -17,6 +17,9 @@ COPY public /public
 COPY swagger /swagger
 
 WORKDIR docs
+
+ARG REMOVE_NPQ_REFERENCES=false
+ENV REMOVE_NPQ_REFERENCES=${REMOVE_NPQ_REFERENCES}
 RUN bundle exec middleman build --build-dir=../public/api-reference
 
 # Stage 1: Download gems and node modules.
